Intialize beans eagerly. Move some profile tests to unit tests instead
We now initialize spring beans eagerly to speed up query response time
(for that first query that hits the JVM), so we always start with a hot
JVM. This is useful for production environments where it is desired a
responsive JVM when running queries.

The profile group in automation is expensive because it restarts PXF
after every test, and in some cases before the test starts to setup
profile files. Some of the testing that we have can be moved to unit
tests and we can also optimize when PXF needs to restart after every
test. In my local, I am able to run the tests in under 60 seconds vs
under 9 minutes previously.